requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
flask_swagger_ui==4.11.1
flask_mail==0.10.0
mysql-connector-python==8.4.0
jellyfish==1.0.4
bcrypt==3.2.0
Requests==2.32.3
